[ISSUES-1] couchdb::source fails on Ubuntu 12.x
For Ubuntu 12.x and greater xulrunner-dev is not added to dev_pkgs since
this package was removed from the repositories
(http://askubuntu.com/questions/125980/how-do-i-install-xulrunner-in-12-04).
Instead xulrunner is manually downloaded from mozillia and installed to
the correct places on the box. This includes the runtime and source
files (sdk). 

Finally the default couchdb local.ini that comes with the couchdb
source is overwritten using the local.ini template found in the cookbook
so that the [couch_db][bind_address] option can also be exposed when
building from source.